Improve input/output stream handling for CLISP
When neither of both keyword arguments was supplied to (start) it would
fail to return the started process. This is now fixed.

CLISP also does not support binding the stdin/stdout of the new process
to existing streams (as e.g. SBCL does). This is part of the general
API, so an explicit check is added to (run) and (start) that raises an
error if custom streams are supplied under CLISP. This breaks the API,
but an error was being generated anyway (by CLISP), this just makes it
clearer why it does not work.