Avoid redefinitions with cmucl
cmucl.lisp does not place any constraints on the arguments it uses
with defmethod, leading to a redefinition rather than an overload.
Since we know upfront which methods cmucl defines, we can do away with
the redefinition by not providing a default implementation for cmucl.

Closes #12.

Tested with cmu-21a__21a_unicode_-linux-x86.