Remove spurious quoting of environment values (SBCL)
The old code adds simple quotes around environment values
when passing a custom environment to an external program
in SBCL.  This is wrong, because it does not pass the
intended value of the environment variables.

It seems other parts of the code base rely uses the
Bourne Shell to start external programs, which could
actually have required a quoting of some kind, so
I guess there had been confusion between the two
situations at some point.